third_party/blink/web_tests/svg/dynamic-updates/SVGPatternElement-svgdom-y-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' property of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.y.baseVal.value is 50
PASS patternElement.y.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
